requirements.txt
torch
gradio
pillow
pandas
opencv-python
ultralytics
torchvision